Split up the classes into different files
