+++ lib/HikeCard.dart
@@ -1,0 +1,33 @@
+import 'package:flutter/material.dart'; 
+
+class HikeCard extends StatelessWidget{
+  @override
+  Widget build(BuildContext context){
+    return new Card(
+        child: new Column(
+          mainAxisSize: MainAxisSize.min,
+          children: <Widget>[
+            const ListTile(
+            leading: const Icon(Icons.directions_walk),
+            title: const Text('jade lake'),
+            subtitle: const Text('20 mile Backpacking'),
+            ),
+            new ButtonTheme.bar(
+              child: new ButtonBar(
+                children: <Widget>[
+                  new FlatButton(
+                    child: const Text('DONE'),
+                    onPressed: () {}
+                  ),
+                  new FlatButton(
+                    child: const Text('DELETE'),
+                    onPressed: () {}
+                  ),
+                ],
+              )
+            )
+          ]
+        )
+      );
+  }
+}
